base struct

+++ src/Component/Logger/GlobalLogger/GLogger.php
@@ -1,16 +1,16 @@
 <?php
 declare(strict_types = 1);
 
-namespace Chopper\Tools\GLogger;
+namespace Chopper\Component\Logger\GlobalLogger;
 
-use Chopper\Exceptions\GLoggerException;
+use Chopper\Component\Logger\GlobalLogger\Exception\GLoggerException;
 use Chopper\Traits\SingletonTrait;
 use Zend\Log\Logger;
 use Zend\Log\LoggerInterface;
 use Zend\Log\Writer\Stream;
 
 /**
- * GLogger
+ * GlobalLogger
  */
 class GLogger
 {
@@ -28,7 +28,9 @@ class GLogger
      */
     public static function configureGlobalLogger(string $logFilePath): void
     {
-        unlink($logFilePath);
+        if (file_exists($logFilePath)) {
+            unlink($logFilePath);
+        }
         static::$logger = new Logger;
         static::$logger->addWriter(new Stream($logFilePath));
     }
@@ -39,7 +41,7 @@ class GLogger
     public static function getLogger(): LoggerInterface
     {
         if (static::$logger === null) {
-            throw new GLoggerException('GLogger did not configured.');
+            throw new GLoggerException('GlobalLogger did not configured.');
         }
 
         return static::$logger;
